+++ components/UI/Navbar.tsx
@@ -114,96 +114,6 @@ export default function Navbar({ enableTransition }) {
     </>
   );
 
-  // const AccountCard = () => (
-  //   <>
-  //     <Modal
-  //       onClose={onCloseModal}
-  //       isOpen={isOpenModal}
-  //       isCentered
-  //       motionPreset="slideInBottom"
-  //       size={"sm"}
-  //     >
-  //       <ModalOverlay bg="none" backdropFilter="auto" backdropBlur="5px" />
-  //       <ModalContent>
-  //         <ModalHeader />
-  //         <ModalCloseButton />
-  //         <ModalBody>
-  //           <Center>
-  //             <VStack spacing={5}>
-  //               <Avatar
-  //                 src={userData.avatar}
-  //                 rounded="full"
-  //                 size="2xl"
-  //               />
-  //               <VStack>
-  //                 <Text fontSize="26px" fontWeight="bold">
-  //                   {user.user_metadata.username}#{user.user_metadata.tag}
-  //                 </Text>
-  //                 <Text fontSize="14px">{user.email}</Text>
-  //               </VStack>
-  //             </VStack>
-  //           </Center>
-  //           <Center>
-  //             <SimpleGrid
-  //               columns={2}
-  //               spacing={5}
-  //               mt={{ base: "6vw", sm: "4vw", md: "2vw" }}
-  //             >
-  //               <NextLink href="/profile" passHref>
-  //                 <Button as="a" variant="solid" fontSize="16px">
-  //                   Profile
-  //                 </Button>
-  //               </NextLink>
-
-  //               <NextLink href="/account" passHref>
-  //                 <Button as="a" variant="solid" fontSize="16px">
-  //                   Account
-  //                 </Button>
-  //               </NextLink>
-
-
-  //               {userData.pixel ? (
-  //                 <>
-  //                   <NextLink href="/feed" passHref>
-  //                     <Button as="a" variant="solid" fontSize="16px">
-  //                       Feed
-  //                     </Button>
-  //                   </NextLink>
-  //                   <NextLink href="/dashboard" passHref>
-  //                     <Button as="a" variant="solid" fontSize="16px">
-  //                       Dashboard
-  //                     </Button>
-  //                   </NextLink>
-  //                 </>
-  //               ) : (
-  //                 <>
-  //                   <NextLink href="/pixels" passHref>
-  //                     <Button as="a" variant="solid" fontSize="16px">
-  //                       Pixel
-  //                     </Button>
-  //                   </NextLink>
-  //                 </>
-  //               )}
-
-  //               <Button
-  //                 as="a"
-  //                 variant="solid"
-  //                 fontSize="16px"
-  //                 onClick={() => supabase.auth.signOut()}
-  //               >
-  //                 Log out
-  //               </Button>
-  //             </SimpleGrid>
-  //           </Center>
-  //         </ModalBody>
-  //         <Center>
-  //           <ModalFooter />
-  //         </Center>
-  //       </ModalContent>
-  //     </Modal>
-  //   </>
-  // );
-
   return (
     <Flex
       zIndex={"99"}
